make_vfat.py: use ftruncate to create the output file
This option is much better when creating the output file and it also fixes the
issue of the output file being too small for the partition.

make_vfat.py: limit size of copied blocks to 64 MB

As we can have large chunks on vfat images, whenever they are properly
defracted, it makes sense to limit the size of blocks read and copied.

This ensures that huge files do not eat a lot of RAM when being read.